Convert cnetextract and update tests
